+++ src/components/BookingPanel/BookingPanel.example.js
@@ -8,7 +8,7 @@ export const Default = {
   component: BookingPanel,
   props: {
     className: css.example,
-    listing: createListing('listing_1'),
+    listing: createListing('listing_1', { availabilityPlan: { timezone: 'Etc/UTC' } }),
     onSubmit: values => console.log('Submit:', values),
     title: <span>Booking title</span>,
     subTitle: 'Hosted by Author N',
@@ -21,7 +21,10 @@ export const WithClosedListing = {
   component: BookingPanel,
   props: {
     className: css.example,
-    listing: createListing('listing_1', { state: LISTING_STATE_CLOSED }),
+    listing: createListing('listing_1', {
+      availabilityPlan: { timezone: 'Etc/UTC' },
+      state: LISTING_STATE_CLOSED,
+    }),
     onSubmit: values => console.log('Submit:', values),
     title: <span>Booking title</span>,
     subTitle: 'Hosted by Author N',