+++ src/components/FieldSelect/FieldSelect.js
@@ -7,7 +7,17 @@ import { ValidationError } from '../../components';
 import css from './FieldSelect.css';
 
 const FieldSelectComponent = props => {
-  const { rootClassName, className, id, label, input, meta, children, ...rest } = props;
+  const {
+    rootClassName,
+    className,
+    selectClassName,
+    id,
+    label,
+    input,
+    meta,
+    children,
+    ...rest
+  } = props;
 
   if (label && !id) {
     throw new Error('id required when a label is given');
@@ -19,7 +29,7 @@ const FieldSelectComponent = props => {
   // field has been touched and the validation has failed.
   const hasError = touched && invalid && error;
 
-  const selectClasses = classNames(css.select, {
+  const selectClasses = classNames(selectClassName, css.select, {
     [css.selectSuccess]: valid,
     [css.selectError]: hasError,
   });
@@ -38,6 +48,7 @@ const FieldSelectComponent = props => {
 FieldSelectComponent.defaultProps = {
   rootClassName: null,
   className: null,
+  selectClassName: null,
   id: null,
   label: null,
   children: null,
@@ -48,6 +59,7 @@ const { string, object, node } = PropTypes;
 FieldSelectComponent.propTypes = {
   rootClassName: string,
   className: string,
+  selectClassName: string,
 
   // Label is optional, but if it is given, an id is also required so
   // the label can reference the input in the `for` attribute