+++ src/examples.js
@@ -72,7 +72,8 @@ import * as UserCard from './components/UserCard/UserCard.example';
 import * as UserDisplayName from './components/UserDisplayName/UserDisplayName.example';
 
 // forms
-import * as BookingDatesForm from './forms/BookingDatesForm/BookingDatesForm.example';
+import * as BookingTimeForm from './forms/BookingTimeForm/BookingTimeForm.example';
+import * as FieldDateAndTimeInput from './forms/BookingTimeForm/FieldDateAndTimeInput.example';
 import * as EditListingAvailabilityForm from './forms/EditListingAvailabilityForm/EditListingAvailabilityForm.example';
 import * as EditListingDescriptionForm from './forms/EditListingDescriptionForm/EditListingDescriptionForm.example';
 import * as EditListingFeaturesForm from './forms/EditListingFeaturesForm/EditListingFeaturesForm.example';
@@ -102,7 +103,7 @@ export {
   Avatar,
   BookingBreakdown,
   BookingDateRangeFilter,
-  BookingDatesForm,
+  BookingTimeForm,
   BookingTimeInfo,
   BookingPanel,
   Button,
@@ -124,6 +125,7 @@ export {
   FieldCurrencyInput,
   FieldDateRangeController,
   FieldDateInput,
+  FieldDateAndTimeInput,
   FieldDateRangeInput,
   FieldPhoneNumberInput,
   FieldRadioButton,