+++ src/forms/index.js
@@ -1,4 +1,4 @@
-export { default as BookingDatesForm } from './BookingDatesForm/BookingDatesForm';
+export { default as BookingTimeForm } from './BookingTimeForm/BookingTimeForm';
 export { default as ContactDetailsForm } from './ContactDetailsForm/ContactDetailsForm';
 export { default as EditListingAvailabilityForm } from './EditListingAvailabilityForm/EditListingAvailabilityForm';
 export { default as EditListingDescriptionForm } from './EditListingDescriptionForm/EditListingDescriptionForm';